+++ CHANGELOG.md
@@ -22,6 +22,7 @@ A big release with several breaking changes, not all of which can be listed belo
 - `Wgit::Utils.pprint(display:)` param to turn off all logs easily (by setting from ENV etc).
 - `Wgit::Crawler#crawl_site(max_pages:)` param to quit the crawl early.
 - `Wgit::Utils.pprint_search_results(include_score:)` param to print the score of each result.
+- `Wgit::Utils.pprint_all_search_results` method to print all matches for each result.
 ### Changed/Removed
 - Update Wgit to work with ruby v3 and above, removed support for older versions.
 - Removed default extractors `meta_robots` and `meta_wgit` without losing any functionality.